Final version

+++ src/main/java/inter/venture/project/domain/user/service/UserService.java
@@ -1,11 +1,14 @@
 package inter.venture.project.domain.user.service;
 
+import inter.venture.project.domain.user.dto.publicDto.UserDto;
+import inter.venture.project.domain.user.dto.privateDto.UserDtoPrivate;
+import inter.venture.project.domain.user.entity.Device;
 import inter.venture.project.domain.user.entity.User;
+import inter.venture.project.domain.user.mapper.UserMapper;
+import inter.venture.project.domain.user.repository.DeviceRepository;
 import inter.venture.project.domain.user.repository.UserRepository;
-import inter.venture.project.domain.user.request.CreateUserRequest;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -19,45 +22,62 @@ public class UserService {
     private UserRepository userRepository;
 
     @Autowired
+    private DeviceRepository deviceRepository;
+
+    @Autowired
     PasswordEncoder passwordEncoder;
 
 
-    public List<User> list(){
-        return userRepository.findAll();
+    public List<UserDto> list(){
+        return UserMapper.instance.listOfUsersToListOfUserDto(userRepository.findAll());
     }
 
-    public User create(CreateUserRequest createUserRequest) throws ValidationException {
-        String username = createUserRequest.getUsername();
-        if (this.userRepository.existsByUsername(username)){
+    public UserDto create(UserDtoPrivate userDtoPrivate) throws ValidationException {
+        User user = UserMapper.instance.userDtoPrivateToUser(userDtoPrivate);
+
+        if (this.userRepository.existsByUsername(user.getUsername())){
             throw new IllegalArgumentException("The given username is already taken.");
         }
-        String encodedPassword = new BCryptPasswordEncoder().encode(createUserRequest.getPassword());
-        String firstName = createUserRequest.getFirstName();
-        String lastName = createUserRequest.getLastName();
+        //Password is given in plain text, but we want to bcrypt it and then sent it to the DB encrypted
+        String encodedPassword = passwordEncoder.encode(user.getPassword());
+        user.setPassword(encodedPassword);
 
-        return this.userRepository.save(new User(username, encodedPassword, firstName, lastName));
+        return UserMapper.instance.userToUserDto(this.userRepository.save(user));
     }
 
 
-    public User get(Long id) {
-        return userRepository.getOne(id);
+    public UserDto get(Long id) {
+        return UserMapper.instance.userToUserDto(userRepository.getOne(id));
     }
 
-        public User update(Long id, User user) {
+    public UserDto update(Long id, UserDtoPrivate userDtoPrivate) {
         User existingUser = userRepository.getOne(id);
+        User user = UserMapper.instance.userDtoPrivateToUser(userDtoPrivate);
         //Password was changed, encryption needs to be done before sending back to the database
+        //Checking if the passwords are the same, if not then change it and bcrypt it
         if(!existingUser.getPassword().equals(passwordEncoder.encode(user.getPassword()))){
             user.setPassword(passwordEncoder.encode(user.getPassword()));
         }
+        //Compares the User from the DB and the existing one and updates it accordingly
         BeanUtils.copyProperties(user, existingUser, "id");
-        return userRepository.saveAndFlush(existingUser);
+        return UserMapper.instance.userToUserDto(userRepository.saveAndFlush(existingUser));
     }
-
 
 
     public void delete(Long id) {
         //Also need to check for children record before deleting
-        userRepository.deleteById(id);
+        if(userRepository.findById(id).isPresent()) {
+            User user = userRepository.findById(id).get();
+
+            //Delete all devices for the user that we are deleting
+            List<Device> devices = deviceRepository.listUserDevices(user.getUsername());
+            if (!devices.isEmpty()) {
+                for (Device device : devices) {
+                    deviceRepository.delete(device);
+                }
+            }
+            userRepository.deleteById(id);
+        }
     }
